+++ source/Nuke.Common/CI/WoodpeckerCI/Configuration/WoodpeckerCIRunStep.cs
@@ -11,7 +11,7 @@ public class WoodpeckerCIRunStep : WoodpeckerCIStep
 {
     public override required string Name { get; set; }
 
-    public override string DockerImage { get; set; } = DotNetDockerImages.Current;
+    public override string DockerImage { get; set; } = DotNetDockerImages.ForCurrentRuntime;
 
     public IEnumerable<string> InvokedTargets { get; set; }